Added Some Pictures for my Hexo Blog

+++ js/Redirect-to-real-site.js
@@ -2,51 +2,35 @@
 // Fucking fake website that use my source code to deploy yours
 // 去你的直接拿我的网站源码去部署你们所谓的**自己的站点**
 var domain = document.domain;
-switch (domain) {
-  case "bili33.top": {
-    undefined;
-    break;
-  }
-  case "127.0.0.1": {
-    undefined;
-    break;
-  }
-  case "localhost": {
-    undefined;
-    break;
-  }
-  case "blog.ninym.top": {
-    if (localStorage.getItem("Mirror-Agreement") == null) {
-      usemirror = confirm(
-        "你正在访问镜像站，只有主站不可用的时候我们才推荐你访问镜像站\n按确定继续访问，或者按取消返回主站\n（确定后此提示不再弹出）"
-      );
-      if (usemirror) {
-        localStorage.setItem("Mirror-Agreement", "true");
+if (
+  domain != "bili33.top" &&
+  domain != "blog.ninym.top" &&
+  domain != "127.0.0.1" &&
+  domain != "localhost" &&
+  domain != "webcache.googleusercontent.com"
+) {
+  // webcache.googleusercontent.com by Google Cache
+  alert(
+    "看起来您正在访问的网页 " +
+      window.location.href +
+      " 并非官方网页，有可能您正在访问Demo站，我们将为您跳转到原网站。"
+  );
+  var pathname = window.location.pathname;
+  window.location = "http://bili33.top" + pathname;
+} else {
+  switch (domain) {
+    case "blog.ninym.top": {
+      if (localStorage.getItem("Mirror-Agreement") == null) {
+        usemirror = confirm(
+          "你正在访问镜像站，只有主站不可用的时候我们才推荐你访问镜像站\n按确定继续访问，或者按取消返回主站\n（确定后此提示不再弹出）"
+        );
+        if (usemirror) {
+          localStorage.setItem("Mirror-Agreement", "true");
+        }
+      } else {
+        break;
       }
-    } else {
       break;
     }
-    break;
   }
-
-  case "webcache.googleusercontent.com": {
-    usecached = confirm(
-      "It seems that you are accessing a cached page. Do you want to redirect to the real-time site?"
-    );
-    if (usecached) {
-      window.location.href = "http://bili33.top/";
-    } else {
-      undefined;
-    }
-    break;
-  }
-  default:
-    alert(
-      "看起来您正在访问的网页 " +
-        window.location.href +
-        " 并非官方网页，有可能您正在访问Demo站，我们将为您跳转到原网站。"
-    );
-    var pathname = window.location.pathname;
-    window.location = "http://bili33.top" + pathname;
-    break;
 }
